clean up github cloudquery plugin
