+++ connectors/github-resolve-repos-cq/src/main.ts
@@ -3,6 +3,8 @@ import { createServeCommand } from "@cloudquery/plugin-sdk-javascript/plugin/ser
 import { newGithubResolveReposPlugin } from "./plugin.js";
 
 const main = () => {
+  // Interestingly this isn't actually a promise but it's not being interpreted correctly
+  // eslint-disable-next-line @typescript-eslint/no-floating-promises
   createServeCommand(newGithubResolveReposPlugin()).parse();
 };
