--- connectors/github-resolve-repos-cq/src/streams/index.ts
@@ -1,2 +1,0 @@
-import { Repositories } from "./repositories.js";
-export { Repositories };
